+++ src/components/@common/Modal/UserConfigModal/index.tsx
@@ -1,8 +1,7 @@
-import useConfirmModal from '@/hooks/useConfirmModal';
-import { useCanWithdrawl } from '@/queries/Auth/useCanWithdrawal';
+import useWithdrawal from '@/hooks/User/useWithdrawal';
 import { userState } from '@/store/userState';
-import { FC, useState } from 'react';
-import { useRecoilValue } from 'recoil';
+import { FC } from 'react';
+import { useRecoilState } from 'recoil';
 import Modal from '..';
 import * as Style from './styles';
 
@@ -11,18 +10,9 @@ interface UserConfigModalProps {
 }
 
 const UserConfigModal: FC<UserConfigModalProps> = ({ handleModal }) => {
-  const [trigger, setTrigger] = useState(false);
-  const userData = useRecoilValue(userState);
-  const { openConfirmModal, closeConfirmModal } = useConfirmModal();
+  const [userData] = useRecoilState(userState);
 
-  const handleWithdrawalModal = () => {
-    openConfirmModal({ type: 'WITHDRAWAL_HAS_ADMIN', confirm: closeConfirmModal });
-  };
-  const { data } = useCanWithdrawl({ modalHandler: handleWithdrawalModal, trigger });
-
-  const onClickWithdrawal = () => {
-    setTrigger(true);
-  };
+  const { onClickWithdrawal } = useWithdrawal();
 
   return (
     <Modal.Frame width="493px" height="284px" onClick={handleModal}>
